improved zsh alias

+++ bin/copy_configs.sh
@@ -10,16 +10,18 @@ TARGET_DIR="$HOME/.config"
 
 mkdir -p "$TARGET_DIR"
 
-echo "Copying configuration files from $CONFIGS_DIR to $TARGET_DIR"
-for folder in "$CONFIGS_DIR"/*/; do
-    folder_name=$(basename "$folder")
-    target_folder="$TARGET_DIR/$folder_name"
-    if [ -d "$target_folder" ]; then
-        echo "Replacing existing folder: $target_folder"
-        rm -rf "$target_folder"
+echo "Copying configuration folders from $CONFIGS_DIR to $TARGET_DIR"
+for folder in "$CONFIGS_DIR"/*; do
+    if [ -d "$folder" ]; then
+        folder_name=$(basename "$folder")
+        target_folder="$TARGET_DIR/$folder_name"
+        if [ -d "$target_folder" ]; then
+            echo "Replacing existing folder: $target_folder"
+            rm -rf "$target_folder"
+        fi
+        cp -r "$folder" "$target_folder"
+        echo "Copied $folder_name"
     fi
-    cp -r "$folder" "$target_folder"
-    echo "Copied $folder_name"
 done
 
 if [ -f "$CONFIGS_DIR/zshrc" ]; then
